all: remove the use of net.Buffers{}
I'm a bit unsure here. The OS is supposed to use writev on these
and make it look like a single tranmission. However I'm not sure
whether does is guaranteed? By default? Also on Windows?

Maybe it that fails we end up with header or payload missing? Got
w bug report for Windows where I suspect this to happen..
eventually?